+++ tests/foreign_keys/test_many_to_many_field_new_naming_multiple.py
@@ -33,11 +33,6 @@ class Studio(edgy.StrictModel):
         registry = models
 
 
-def test_check_tablename():
-    assert Studio.meta.fields["users"].through.meta.tablename == "studiousersthrough"
-    assert Studio.meta.fields["admins"].through.meta.tablename == "studioadminsthrough"
-
-
 @pytest.fixture(autouse=True, scope="function")
 async def create_test_database():
     async with database:
@@ -45,6 +40,11 @@ async def create_test_database():
         yield
         if not database.drop:
             await models.drop_all()
+
+
+async def test_check_tablename():
+    assert Studio.meta.fields["users"].through.meta.tablename == "studiousersthrough"
+    assert Studio.meta.fields["admins"].through.meta.tablename == "studioadminsthrough"
 
 
 async def test_many_to_many_many_fields():